canon-cups-ufr2: Rework build script with RPM spec
Strongly inspired by the AUR package [1], this commit removes most
manual build and install commands in favor of reusing the RPM spec file.

With the rework and a small patch, cngplp now builds and is installed
correctly. Previously, we would skip building cngplp, resulting in
broken symlinks in the output folder.

[1]: https://aur.archlinux.org/packages/cnrdrvcups-lb

Fixes #380572